+++ python/tablers/typing.py
@@ -70,7 +70,7 @@ class TfSettingItems(TypedDict, total=False):
     text_expand_ligatures: bool
     explicit_h_edges: list[Edge] | None
     explicit_v_edges: list[Edge] | None
-    exclude_white_edges: bool
+    exclude_background_colored_edges: bool
 
 
 class WordsExtractSettingsItems(TypedDict, total=False):